+++ pkg/apis/serving/custom_builder.go
@@ -168,6 +168,18 @@ func (b *CustomServingJobBuilder) Annotations(annotations map[string]string) *Cu
 	return b
 }
 
+// Labels is used to add labels for job
+func (b *CustomServingJobBuilder) Labels(labels map[string]string) *CustomServingJobBuilder {
+	if labels != nil && len(labels) != 0 {
+		s := []string{}
+		for key, value := range labels {
+			s = append(s, fmt.Sprintf("%v=%v", key, value))
+		}
+		b.argValues["label"] = &s
+	}
+	return b
+}
+
 // Datas is used to mount k8s pvc to job pods,match option --data
 func (b *CustomServingJobBuilder) Datas(volumes map[string]string) *CustomServingJobBuilder {
 	if volumes != nil && len(volumes) != 0 {