+++ pkg/apis/serving/tensorflow_builder.go
@@ -188,6 +188,18 @@ func (b *TFServingJobBuilder) Annotations(annotations map[string]string) *TFServ
 	return b
 }
 
+// Labels is used to add labels for job
+func (b *TFServingJobBuilder) Labels(labels map[string]string) *TFServingJobBuilder {
+	if labels != nil && len(labels) != 0 {
+		s := []string{}
+		for key, value := range labels {
+			s = append(s, fmt.Sprintf("%v=%v", key, value))
+		}
+		b.argValues["label"] = &s
+	}
+	return b
+}
+
 // Datas is used to mount k8s pvc to job pods,match option --data
 func (b *TFServingJobBuilder) Datas(volumes map[string]string) *TFServingJobBuilder {
 	if volumes != nil && len(volumes) != 0 {