+++ pkg/apis/training/mpijob_builder.go
@@ -119,6 +119,18 @@ func (b *MPIJobBuilder) Annotations(annotations map[string]string) *MPIJobBuilde
 	return b
 }
 
+// Labels is used to add labels for job
+func (b *MPIJobBuilder) Labels(labels map[string]string) *MPIJobBuilder {
+	if labels != nil && len(labels) != 0 {
+		s := []string{}
+		for key, value := range labels {
+			s = append(s, fmt.Sprintf("%v=%v", key, value))
+		}
+		b.argValues["label"] = &s
+	}
+	return b
+}
+
 // Datas is used to mount k8s pvc to job pods,match option --data
 func (b *MPIJobBuilder) Datas(volumes map[string]string) *MPIJobBuilder {
 	if volumes != nil && len(volumes) != 0 {