+++ pkg/apis/types/submit.go
@@ -76,6 +76,9 @@ type CommonSubmitArgs struct {
 	// Annotations defines pod annotations of job,match option --annotation
 	Annotations map[string]string `yaml:"annotations"`
 
+	// Labels specify the job labels and it is work for pods
+	Labels map[string]string `yaml:"labels"`
+
 	// IsNonRoot is root user or not
 	IsNonRoot bool `yaml:"isNonRoot"`