+++ pkg/apis/types/submit_sparkjob.go
@@ -9,6 +9,10 @@ type SubmitSparkJobArgs struct {
 	Jar          string          `yaml:"Jar"`
 	Executor     *Executor       `yaml:"Executor"`
 	Driver       *Driver         `yaml:"Driver"`
+	// Annotations defines pod annotations of job,match option --annotation
+	Annotations map[string]string `yaml:"annotations"`
+	// Labels specify the job labels and it is work for pods
+	Labels map[string]string `yaml:"labels"`
 }
 
 type Driver struct {