+++ pkg/training/submit_etjob.go
@@ -44,6 +44,9 @@ func SubmitETJob(namespace string, submitArgs *types.SubmitETJobArgs) (err error
 	}
 	// if error is unknown,return an error
 	if err != types.ErrTrainingJobNotFound {
+		if err == types.ErrNoPrivilegesToOperateJob {
+			return fmt.Errorf("the job %s is already exist and it owned by other user,you have no privileges to operate it", submitArgs.Name)
+		}
 		return err
 	}
 	// the master is also considered as a worker