+++ pkg/training/submit_volcanojob.go
@@ -23,6 +23,9 @@ func SubmitVolcanoJob(namespace string, submitArgs *types.SubmitVolcanoJobArgs) 
 	}
 	// if error is unknown,return an error
 	if err != types.ErrTrainingJobNotFound {
+		if err == types.ErrNoPrivilegesToOperateJob {
+			return fmt.Errorf("the job %s is already exist and it owned by other user,you have no privileges to operate it", submitArgs.Name)
+		}
 		return err
 	}
 	volcanoChart := util.GetChartsFolder() + "/volcanojob"