+++ charts/etjob/CHANGELOG.md
@@ -2,4 +2,8 @@
 
 * support et operator
 
+### 0.2.0
+
+* support job labels
+