+++ charts/sparkjob/CHANGELOG.md
@@ -7,3 +7,7 @@
 
 * update version to 2.4.5
 
+### 0.3.0
+
+* support job labels and annotations
+